# port BitArray.And, Or, Xor and Not to Vector<T>