@adamsitnik 2022-07-29
I was able to reduce the gap using `Unsafe` APIs, but it's still not acceptable.

```ini
BenchmarkDotNet=v0.13.1.1828-nightly, OS=Windows 11 (10.0.22000.795/21H2)
AMD Ryzen Threadripper PRO 3945WX 12-Cores, 1 CPU, 24 logical and 12 physical cores
.NET SDK=7.0.100-preview.6.22352.1
  [Host]     : .NET 7.0.0 (7.0.22.32404), X64 RyuJIT AVX2
```


|      Method |  Job | Size |       Mean | Ratio |
|------------ |----- |----- |-----------:|------:|
| BitArrayNot |   PR |  512 |  7.6331 ns |  1.32 |
| BitArrayNot | main |  512 |  5.8448 ns |  1.00 |
|             |      |      |            |       |
| BitArrayAnd |   PR |  512 | 12.5956 ns |  1.30 |
| BitArrayAnd | main |  512 |  9.7358 ns |  1.00 |
|             |      |      |            |       |
|  BitArrayOr |   PR |  512 | 12.7886 ns |  1.36 |
|  BitArrayOr | main |  512 |  9.4089 ns |  1.00 |
|             |      |      |            |       |
| BitArrayXor |   PR |  512 | 12.7228 ns |  1.36 |
| BitArrayXor | main |  512 |  9.3926 ns |  1.00 |